Swap to scoped karma-coverage fix. Explanation below...
Unfortunately karma-coverage doesn't (always?) return a return value indicating if the coverage run was successful or not. That means khutzpa's return value won't, for instance, stop a build script when the coverage check fails as echo %ERRORLEVEL% (cmd) or echo $LastExitCode (pwsh) always return 0 indicating success.

Here's an issue on karma-coverage from last December:
* Exit code 0 even if the coverage fails when reporter's file is missing #485
* https://github.com/karma-runner/karma-coverage/issues/485

Luckily there's also a pretty straightforward PR fixing it
* https://github.com/karma-runner/karma-coverage/pull/486
* fix: exit code when reporter file is not provided #486

Unfortunately, though the PR was quickly reviewed & approved by a pass committer to karma-coverage, nobody has merged it.

As a workaround, a scoped version has been set up here with the fix applied:
https://www.npmjs.com/package/@rufwork/karma-coverage

Unfortunately scoped packages don't follow karma's convention for auto-loading plugins.

https://karma-runner.github.io/2.0/config/plugins.html
> By default, Karma loads all sibling NPM modules which have a name starting with `karma-*`.

That means we have to list *all* of our plugins by name in the karma config.

But with those two changes, we seem to be in business.